Move query-related components to subdirectory
