+++ src/web/components/IdeApp.vue
@@ -8,7 +8,7 @@ import IdeSidebar from "./IdeSidebar.vue"
 import IdeHeader from "./IdeHeader.vue"
 import ResultsPane from "./ResultsPane.vue"
 import CodeMirrorEditor from "./CodeMirrorEditor.vue"
-import QueryModal from "./QueryModal.vue"
+import QueryModal from "./query/QueryModal.vue"
 import { provideEditor } from "./editor-context"
 import type { ResultsPaneType } from "./use-results-pane"
 import { ScamperInstance } from "../../scamper-instance"
