+++ src/web/components/query/query-modal-extension.ts
